// hivemq-edge-frontend/src/api/__generated__/models/ValuesTree.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { ObjectNode } from './ObjectNode';

export type ValuesTree = {
    /**
     * List of result items that are returned by this endpoint
     */
    items: Array<ObjectNode>;
};